Refactor Junit5 for BaseTestAESCCM and BaseTestAESCCM2 tests
This change consolidates the tests previously associated with BaseTestAESCCM into the parameterized TestAESCCM class, and the tests from BaseTestAESCCM2 into the parameterized TestAESCCM class.
